Optional directory structure similar to /pentest
Create a directory structure in a given path based on the Kali
aplications menu in Kali Linux. This is useful when you want to browse
the Kali menu inside a terminal or when you use a WM that is not fully
XDG compatible.

menu2dir parses kali-applications.menu using the python-xdg package.
In order for the XDG parse function to detect the DesktopEntry nodes in
kali-applications.menu, the following line had to be added to
kali-applications.menu:

        <AppDir>/usr/share/kali-menu/applications</AppDir>